Release 5.2.0

+++ src/solver/toposort.c
@@ -3,7 +3,7 @@
 //
 //   Project:  EPA SWMM5
 //   Version:  5.2
-//   Date:     03/24/21   (Build 5.2.0)
+//   Date:     11/01/21   (Build 5.2.0)
 //   Author:   L. Rossman
 //
 //   Topological sorting of conveyance network links
@@ -292,7 +292,7 @@ void  findCycles()
     int i;
 
     // --- allocate arrays
-    AdjList  = (int *) calloc(2*Nobjects[LINK], sizeof(int));
+    AdjList  = (int *) calloc(2*(size_t)Nobjects[LINK], sizeof(int));
     StartPos = (int *) calloc(Nobjects[NODE], sizeof(int));
     Stack    = (int *) calloc(Nobjects[NODE], sizeof(int));
     Examined = (char *) calloc(Nobjects[NODE], sizeof(char));
